LayoutTests/imported/w3c/web-platform-tests/html/semantics/scripting-1/the-script-element/module/referrer-policy-for-descendants.sub-expected.txt


FAIL Parent module delivered with `no-referrer-when-downgrade` policy importing a same-origin descendant script. assert_equals: Descendant referrer should be the parent's full URL. expected "http://web-platform.test:8800/html/semantics/scripting-1/the-script-element/module/resources/import-referrer-checker.sub.js?name=same_same&pipe=header(Referrer-Policy,no-referrer-when-downgrade)" but got ""
FAIL Parent module delivered with `origin` policy importing a same-origin descendant script. assert_equals: Descendant referrer should be the parent's origin. expected "http://web-platform.test:8800/" but got ""
FAIL Parent module delivered with `same-origin` policy importing a same-origin descendant script. assert_equals: Descendant referrer should be the parent's full URL. expected "http://web-platform.test:8800/html/semantics/scripting-1/the-script-element/module/resources/import-referrer-checker.sub.js?name=same_remote_so&pipe=header(Referrer-Policy,same-origin)" but got ""
FAIL Parent module delivered with `origin-when-cross-origin` policy importing a cross-origin descendant script. assert_equals: Remote descendant referrer should be the parent's origin. expected "http://web-platform.test:8800/" but got ""
FAIL Remote parent module delivered with `origin-when-cross-origin` policy importing a same-origin-to-parent-module descendant script. assert_equals: Same-origin descendant referrer should be the parent's full URL. expected "http://www1.web-platform.test:8800/html/semantics/scripting-1/the-script-element/module/resources/import-referrer-checker.sub.js?name=remote_same&pipe=header(Referrer-Policy,origin-when-cross-origin)" but got ""
FAIL Remote parent module delivered with `same-origin` policy importing a same-origin-to-parent-module descendant script. assert_equals: Same-origin descendant referrer should be the parent's full URL. expected "http://www1.web-platform.test:8800/html/semantics/scripting-1/the-script-element/module/resources/import-referrer-checker.sub.js?name=remote_same_so&pipe=header(Referrer-Policy,same-origin)" but got ""
FAIL Remote parent module delivered with `origin-when-cross-origin` policy importing a cross-origin-to-parent-module descendant script. assert_equals: Remote-origin descendant referrer should be the parent's origin. expected "http://www1.web-platform.test:8800/" but got ""
PASS Remote parent module delivered with `same-origin` policy importing a cross-origin-to-parent-module descendant script.
PASS Parent module delivered with invalid policy importing a same-origin descendant script.
